gemini_chatbot/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-chroma
pypdf
fastapi
uvicorn[standard]
python-multipart